Create install_flasher
this is a simple enough bash script for installing the flasher on ubuntu. i hope to add checks and some documentation later but i just want to get this up here for now. i am in the process of testing and refining it so i wont push in onto the install instructions until i have tested it on a few laptops.